Link sort fixes
- consistently use org-property-link-sort-mode.  There were a few
instances where this variable was not renamed from
org-insert-property-link-sort-mode.

- default to sort by link name unless org-insert-property-link-sort-mode
is set to "description".  If unset, sort by link.

- Some autocomplete packages perform additional unwanted sorting.  So,
ensure the strings passed will sort in the desired order.  For
sort-by-link, present link first.  For sort-by-description, present
description first.  Make sure the entire string (with both link and
description) is padded to the maximum length.